lxd/instance: Move VMAgentData to instancetype
