+++ lxd-agent/devlxd.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/gorilla/mux"
 	"github.com/lxc/lxd/lxd/daemon"
-	deviceConfig "github.com/lxc/lxd/lxd/device/config"
+	"github.com/lxc/lxd/lxd/instance/instancetype"
 	"github.com/lxc/lxd/lxd/util"
 	"github.com/lxc/lxd/shared"
 	"github.com/lxc/lxd/shared/logger"
@@ -31,14 +31,6 @@ type devLxdResponse struct {
 	content interface{}
 	code    int
 	ctype   string
-}
-
-type instanceData struct {
-	Name        string                         `json:"name"`
-	CloudInitID string                         `json:"cloud_init_id"`
-	Location    string                         `json:"location"`
-	Config      map[string]string              `json:"config,omitempty"`
-	Devices     map[string]deviceConfig.Device `json:"devices,omitempty"`
 }
 
 func okResponse(ct interface{}, ctype string) *devLxdResponse {
@@ -63,7 +55,7 @@ var devlxdConfigGet = devLxdHandler{"/1.0/config", func(d *Daemon, w http.Respon
 		return &devLxdResponse{"internal server error", http.StatusInternalServerError, "raw"}
 	}
 
-	var instance instanceData
+	var instance instancetype.VMAgentData
 
 	err = json.Unmarshal(data, &instance)
 	if err != nil {
@@ -90,7 +82,7 @@ var devlxdConfigKeyGet = devLxdHandler{"/1.0/config/{key}", func(d *Daemon, w ht
 		return &devLxdResponse{"internal server error", http.StatusInternalServerError, "raw"}
 	}
 
-	var instance instanceData
+	var instance instancetype.VMAgentData
 
 	err = json.Unmarshal(data, &instance)
 	if err != nil {
@@ -111,7 +103,7 @@ var devlxdMetadataGet = devLxdHandler{"/1.0/meta-data", func(d *Daemon, w http.R
 		return &devLxdResponse{"internal server error", http.StatusInternalServerError, "raw"}
 	}
 
-	var instance instanceData
+	var instance instancetype.VMAgentData
 
 	err = json.Unmarshal(data, &instance)
 	if err != nil {
@@ -137,7 +129,7 @@ var devlxdAPIGet = devLxdHandler{"/1.0", func(d *Daemon, w http.ResponseWriter, 
 		return &devLxdResponse{"internal server error", http.StatusInternalServerError, "raw"}
 	}
 
-	var instance instanceData
+	var instance instancetype.VMAgentData
 
 	err = json.Unmarshal(data, &instance)
 	if err != nil {
@@ -152,7 +144,7 @@ var devlxdDevicesGet = devLxdHandler{"/1.0/devices", func(d *Daemon, w http.Resp
 		return &devLxdResponse{"internal server error", http.StatusInternalServerError, "raw"}
 	}
 
-	var instance instanceData
+	var instance instancetype.VMAgentData
 
 	err = json.Unmarshal(data, &instance)
 	if err != nil {
